Marker: move source files
