+++ src/ScottPlot/Marker/MarkerShape.cs
@@ -1,0 +1,23 @@
+﻿namespace ScottPlot;
+
+public enum MarkerShape
+{
+    none,
+    filledCircle,
+    filledSquare,
+    openCircle,
+    openSquare,
+    filledDiamond,
+    openDiamond,
+    asterisk,
+    hashTag,
+    cross,
+    eks,
+    verticalBar,
+    triUp,
+    triDown,
+    filledTriangleUp,
+    filledTriangleDown,
+    openTriangleUp,
+    openTriangleDown,
+}
